+++ rust_dev_preview/cross_service/rest_ses/src/client.rs
@@ -1,8 +1,4 @@
-use aws_sdk_rdsdata::{
-    client::fluent_builders::ExecuteStatement,
-    model::{Field, SqlParameter},
-    Client,
-};
+use aws_sdk_rdsdata::{client::fluent_builders::ExecuteStatement, Client};
 use secrecy::{ExposeSecret, Secret};
 
 use crate::configuration::SdkSettings;
@@ -45,13 +41,13 @@ impl Clone for RdsClient {
     }
 }
 
-pub fn param(name: &str, value: String) -> SqlParameter {
-    SqlParameter::builder()
-        .name(name)
-        .value(Field::StringValue(value.clone()))
-        .build()
-}
-
+/// This is a declarative macro that builds a Vec<SqlParameter> to use in ExecuteStatementBuilder::set_parameters.
+/// It requires that all parameters will be sent as strings.
+///
+/// ```
+/// client.execute_statement("INSERT values (:name) INTO table;")
+/// .set_parameters(params![("name"), ("rust")])
+/// ```
 #[macro_export]
 macro_rules! params {
     ($((
@@ -59,10 +55,12 @@ macro_rules! params {
         $value:expr
     )),* ) => {
         {
-            use crate::client::param;
             let mut v = Vec::new();
             $(
-                v.push(param($name, $value.to_owned()));
+                v.push(aws_sdk_rdsdata::model::SqlParameter::builder()
+                    .name($name)
+                    .value(aws_sdk_rdsdata::model::Field::StringValue($value.clone()))
+                    .build());
             )*
             Some(v)
         }
